Add passport authorization middleware, login, rename sessions route file
sessions.js in routes is now called authenticate.js. Auth system files
now live in src/config/auth. POSTing to authenticate will now check a
submitted username and password against the database and determine if a
match for both is found, responding with a token if so, otherwise
telling the user that they have submitted an invalid combination.