Make server respond to authentication requests with token on valid login
POST requests to /authenticate will now be checked for an email and
password that are compared against the database. If a user record with
the given email is found, the hashed password from that record is
compared against a hashed version of the POSTed password and if it
matches a token is returned in the response. If the email is not found
in the user records or the hash password submission does not match the
hashed password from the database, no token is generated or returned.