Remove token secret and replace with environment variable.
Was publishing secret to public repository. Now the code uses dotenv npm
to process the key from a .env file that lives in the project's root
directory (this file is in gitignore, so make your own) that contains a
line that says:
secretOrKey=<INSERT SECRET HERE>
You have to create this file and assign your own secret where the
angle-bracketed phrase is.