Make skip call button not reassign same call. (#151)
Components now call actions more specifically from props to allow for an order of actions that doesn't set the current call to available in the database before asking the db to assign an available call. That behavior was resulting in the newly assigned call being the one that was just released. Now we get a new call before releasing the current one.